refactor(server-remote-standalone): inline auth into server and simplify middleware
Inline auth.ts into server.ts, removing the separate file. Drop the
per-request `c.set('auth', auth)` middleware — the auth guard now
references auth directly from the closure. Use createFactory({ initApp })
for CORS and factory.createMiddleware for the auth guard, mirroring the
cloudflare package's organization.
