+++ packages/server-remote-standalone/src/index.ts
@@ -1,4 +1,3 @@
-export type { StandaloneAuthConfig } from './auth';
-export type { StandaloneHubConfig } from './server';
+export type { StandaloneAuthConfig, StandaloneHubConfig } from './server';
 export { createRemoteHub } from './server';
 export { BunSqliteUpdateLog } from './storage';
